Minor text updates for Readability and Information
To the Welcome Screen, I added a line for the player agreeing to Back Up their data regularly. Prior to this there was no explicit recommendation to do so present in the addon, rules, or the FAQ.

Updated the line in the contract about playing on "multiple computers" to playing on "Different Computers". Currently searching "multiple computers" on the FAQ will return nothing. The Q/A below however uses "different computers" several times. Improves search ability of terms.

I did not color this page in this Pull Request. However I do believe a splash of color couldn't hurt the readability page. Example in this screenshot https://i.imgur.com/7wKEsTb.png

For the Main Menu. Added a line reminding players of mechanical things they should do while running the addon regularly, with a splash of color for readability.

Added descriptions for Dungeon and Death Statistics tabs, and removed an errant, second Achievement description

Refactored a couple of text lines for legibility in code.